refactor: refactoring user modules (#1)
