+++ src/app.module.ts
@@ -4,11 +4,9 @@ import { AppService } from './app.service';
 import { SharedModule } from './shared/shared.module';
 import { CoreModule } from './core/core.module';
 import { UserModule } from './user/user.module';
-import { AuthMiddleware } from './auth/auth.middleware';
 import { PrismaModule } from './prisma/prisma.module';
 import { AuthModule } from './auth/auth.module';
 import { ShoppingListModule } from './shopping-list/shopping-list.module';
-import { UserProfileModule } from './user-profile/user-profile.module';
 import { ShoppingListProductModule } from './shopping-list-product/shopping-list-product.module';
 
 @Module({
@@ -19,7 +17,6 @@ import { ShoppingListProductModule } from './shopping-list-product/shopping-list
     PrismaModule,
     AuthModule,
     ShoppingListModule,
-    UserProfileModule,
     ShoppingListProductModule,
   ],
   controllers: [AppController],
