--- src/user-profile/user-profile.controller.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { UserProfileController } from './user-profile.controller';
-
-describe('UserProfileController', () => {
-  let controller: UserProfileController;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      controllers: [UserProfileController],
-    }).compile();
-
-    controller = module.get<UserProfileController>(UserProfileController);
-  });
-
-  it('should be defined', () => {
-    expect(controller).toBeDefined();
-  });
-});
