--- src/user-profile/user-profile.service.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { UserProfileService } from './user-profile.service';
-
-describe('UserProfileService', () => {
-  let service: UserProfileService;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      providers: [UserProfileService],
-    }).compile();
-
-    service = module.get<UserProfileService>(UserProfileService);
-  });
-
-  it('should be defined', () => {
-    expect(service).toBeDefined();
-  });
-});
