--- src/user/dto/query-user-filter.dto.ts
@@ -1,13 +1,0 @@
-import { IsOptional, IsString } from 'class-validator';
-import { Transform } from 'class-transformer';
-
-export class QueryUserFilterDto {
-  @IsOptional()
-  @IsString({ message: 'O filtro deve ser uma string válida.' })
-  @Transform(({ value }) => value?.trim())
-  filter?: string;
-
-  @IsOptional()
-  @Transform(({ value }) => parseInt(value, 10))
-  page?: number;
-}
